+++ lsp-types/src/Language/LSP/Protocol/Message/LspId.hs
@@ -6,10 +6,12 @@ import           Data.Hashable
 import           Data.IxMap
 import           Data.Text                          (Text)
 import           GHC.Generics
+import           Prettyprinter
 
 import           Language.LSP.Protocol.Types.Common
 import           Language.LSP.Protocol.Internal.Method
 import           Language.LSP.Protocol.Message.Meta
+import           Language.LSP.Protocol.Utils.Misc
 
 -- | Id used for a request, Can be either a String or an Int
 data LspId (m :: Method f Request) = IdInt !Int32 | IdString !Text
@@ -23,6 +25,8 @@ instance A.FromJSON (LspId m) where
   parseJSON v@(A.Number _) = IdInt <$> A.parseJSON v
   parseJSON  (A.String  s) = return (IdString s)
   parseJSON _              = fail "LspId"
+
+deriving via ViaJSON (LspId m) instance Pretty (LspId m)
 
 instance IxOrd LspId where
   type Base LspId = SomeLspId